requirements.txt
Telethon==1.38.1
pytz==2024.2
loguru==0.7.3
apache-airflow
python-dotenv== 1.0.1
google-api-python-client==2.156.0
boto3==1.35.85